Don't remount if the container is already running
This change fixes double-mounting when running an exec or a cp on a
running container with the containerd integration enabled. Sharing
layers with overlayfs results in an undefined state and things start to
randomly break with a "No such file or directory" error.

Graph drivers use reference counting and only mount things once, we
can't do this with containerd because we don't know where it mounted the
container rootfs. Instead of reference counting we return the
`/proc/<pid>/root` directory for exec and/or cp to read from.

+++ daemon/containerd/mount.go
@@ -4,11 +4,14 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/containerd/containerd/mount"
 	"github.com/docker/docker/container"
 	"github.com/sirupsen/logrus"
 )
+
+const rootfsMountSuffix = "_rootfs-mount"
 
 // Mount mounts the container filesystem in a temporary location, use defer imageService.Unmount
 // to unmount the filesystem when calling this
@@ -19,15 +22,21 @@ func (i *ImageService) Mount(ctx context.Context, container *container.Container
 		return err
 	}
 
-	// The temporary location will be under /var/lib/docker/... because
-	// we set the `TMPDIR`
-	root, err := os.MkdirTemp("", fmt.Sprintf("%s_rootfs-mount", container.ID))
-	if err != nil {
-		return fmt.Errorf("failed to create temp dir: %w", err)
-	}
+	root := ""
+	if container.State != nil && !container.State.Running && container.Pid == 0 {
+		// The temporary location will be under /var/lib/docker/... because
+		// we set the `TMPDIR`
+		var err error
+		root, err = os.MkdirTemp("", container.ID+rootfsMountSuffix)
+		if err != nil {
+			return fmt.Errorf("failed to create temp dir: %w", err)
+		}
 
-	if err := mount.All(mounts, root); err != nil {
-		return fmt.Errorf("failed to mount %s: %w", root, err)
+		if err := mount.All(mounts, root); err != nil {
+			return fmt.Errorf("failed to mount %s: %w", root, err)
+		}
+	} else {
+		root = fmt.Sprintf("/proc/%d/root", container.Pid)
 	}
 
 	container.BaseFS = root
@@ -38,12 +47,14 @@ func (i *ImageService) Mount(ctx context.Context, container *container.Container
 func (i *ImageService) Unmount(ctx context.Context, container *container.Container) error {
 	root := container.BaseFS
 
-	if err := mount.UnmountAll(root, 0); err != nil {
-		return fmt.Errorf("failed to unmount %s: %w", root, err)
-	}
+	if strings.HasSuffix(root, rootfsMountSuffix) {
+		if err := mount.UnmountAll(root, 0); err != nil {
+			return fmt.Errorf("failed to unmount %s: %w", root, err)
+		}
 
-	if err := os.Remove(root); err != nil {
-		logrus.WithError(err).WithField("dir", root).Error("failed to remove mount temp dir")
+		if err := os.Remove(root); err != nil {
+			logrus.WithError(err).WithField("dir", root).Error("failed to remove mount temp dir")
+		}
 	}
 
 	container.BaseFS = ""
